New methods to load a SQLite extension
This commit adds a new method 'load_extension' to the NativeDB class
to load a SQLite extension. The method takes a first String parameter,
which is the path to the extension file, and a second one, which is the
entry point of the extension. The entry point is optional, and if not
specified, the default entry point 'sqlite3_extension_init' is used.

The corresponding JNI method is also implemented in the NativeDB.c file.